split main.py into appropriate routes files
